requirements.txt
asgiref==3.3.4
Django==3.2
django-debug-toolbar==3.2.1
django-glrm==1.1.3
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1

